ci: run tests in parallel
Move cross-file test helpers into an autoloaded file so each paratest
worker can resolve them.
