+++ src/cms/composer.json
@@ -75,7 +75,10 @@
     "autoload-dev": {
         "psr-4": {
             "Tests\\": "tests/"
-        }
+        },
+        "files": [
+            "tests/Helpers/TransferHelpers.php"
+        ]
     },
     "scripts": {
         "post-autoload-dump": [
