+++ src/cms/tests/Feature/Transfer/TransferRoundTripTest.php
@@ -2,66 +2,13 @@
 
 declare(strict_types=1);
 
-use App\Models\Avg\AvgGoal;
 use App\Models\Avg\AvgResponsibleProcessingRecord;
 use App\Models\Avg\AvgResponsibleProcessingRecordService;
 use App\Models\Organisation;
 use App\Models\Processor;
-use App\Models\Tag;
 use App\Models\User;
-use App\Transfer\Export\BundleExporter;
 use App\Transfer\Import\BundleImporter;
-use App\Transfer\TransferEntityType;
 use Illuminate\Support\Facades\Storage;
-
-function createExportedBundle(Organisation $sourceOrganisation, User $user): array
-{
-    $service = AvgResponsibleProcessingRecordService::factory()
-        ->for($sourceOrganisation)
-        ->create(['name' => 'Dienst X']);
-
-    $record = AvgResponsibleProcessingRecord::factory()
-        ->for($sourceOrganisation)
-        ->create([
-            'name' => 'Verwerking A',
-            'avg_responsible_processing_record_service_id' => $service->id,
-            // the record observer deletes attached processors/systems when these flags are false
-            'has_processors' => true,
-            'has_systems' => true,
-        ]);
-
-    $processor = Processor::factory()->for($sourceOrganisation)->create(['name' => 'Verwerker 1']);
-    $processor->address()->create(['city' => 'Amsterdam']);
-
-    $tag = Tag::factory()->for($sourceOrganisation)->create(['name' => 'Label 1']);
-    $avgGoal = AvgGoal::factory()->for($sourceOrganisation)->create(['goal' => 'Doel 1']);
-
-    $record->processors()->attach($processor);
-    $record->tags()->attach($tag);
-    $record->avgGoals()->attach($avgGoal);
-    $record->fgRemark()->create(['body' => 'FG opmerking']);
-    $record->remarks()->create(['body' => 'Notitie', 'user_id' => $user->id]);
-
-    $path = app(BundleExporter::class)->export(
-        TransferEntityType::AVG_RESPONSIBLE_PROCESSING_RECORD,
-        [$record->id->toString()],
-        [
-            'processors' => [$processor->id->toString()],
-            'tags' => [$tag->id->toString()],
-            'avgGoals' => [$avgGoal->id->toString()],
-        ],
-        $sourceOrganisation,
-    );
-
-    $plan = [
-        $record->id->toString() => ['selected' => true, 'strategy' => null],
-        $processor->id->toString() => ['selected' => true, 'strategy' => null],
-        $tag->id->toString() => ['selected' => true, 'strategy' => null],
-        $avgGoal->id->toString() => ['selected' => true, 'strategy' => null],
-    ];
-
-    return [$path, $plan, $record, $processor];
-}
 
 it('exports records with related items and imports them into another organisation', function (): void {
     Storage::fake('filament');
